+++ frontend/js/router/router.js
@@ -9,6 +9,7 @@ import { ServicesSkeleton } from "../modules/services/components/ServicesSkeleto
 import { ClientsPage } from "../modules/admin/clients/clients.page.js";
 import { ProfessionalsAdminPage } from "../modules/admin/professionals/professionals.page.js";
 import { ServicesAdminPage } from "../modules/admin/services/services.page.js";
+import { AgendaPage } from "../modules/admin/agenda/agenda.page.js";
 import { ProfessionalsPage } from "../modules/professionals/professionals.page.js";
 import { store } from "../state/store.js";
 import { ROLES } from "../constants/roles.js";
@@ -37,6 +38,7 @@ const routes = {
     "/admin/clients": { page: ClientsPage, roles: [ROLES.ADMIN] },
     "/admin/professionals": { page: ProfessionalsAdminPage, roles: [ROLES.ADMIN] },
     "/admin/services": { page: ServicesAdminPage, roles: [ROLES.ADMIN] },
+    "/admin/agenda": { page: AgendaPage, roles: [ROLES.ADMIN] },
     "/login": { page: LoginPage, guestOnly: true },
     "/register": { page: RegisterPage, guestOnly: true },
 };